% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowprofileLayers.R
\name{format_snowprofileLayers}
\alias{format_snowprofileLayers}
\title{Format snowprofileLayers}
\usage{
format_snowprofileLayers(
  obj,
  target = "all",
  hs = NA,
  validate = TRUE,
  dropNAs = TRUE
)
}
\arguments{
\item{obj}{snowprofileLayers object}

\item{target}{string, indicating which fields are auto-filled ('all', 'height', 'depth', 'thickness', 'none')}

\item{hs}{total snow height (cm) if not deductible from given fields}

\item{validate}{Validate \code{obj} with \link{validate_snowprofileLayers}?}

\item{dropNAs}{Do you want to drop all columns consisting of NAs only?}
}
\value{
copy of obj with auto-filled columns
}
\description{
Calculate missing data.frame columns based on the given ones, if possible.
}
