% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-grainDict.R
\docType{data}
\name{grainDict}
\alias{grainDict}
\title{A data.frame storing the grain type colours}
\format{
A data.frame
}
\usage{
grainDict
}
\description{
The colours can be changed by calling the function \link{setColoursGrainType}, see examples below.
}
\examples{

print(grainDict)

## change colours for subsequent plots:
grainDict <- setColoursGrainType('sarp-reduced')

}
\keyword{grainDict}
