% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unobservedBasalLayer.R
\name{hasUnobservedBasalLayer}
\alias{hasUnobservedBasalLayer}
\title{Check whether a profile is observed down to ground or not}
\usage{
hasUnobservedBasalLayer(x)
}
\arguments{
\item{x}{a \link{snowprofile}, or \link{snowprofileLayers} object}
}
\value{
boolean TRUE/FALSE
}
\description{
Check whether a profile is observed down to ground or not
}
