% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowprofile.R
\name{is.snowprofile}
\alias{is.snowprofile}
\title{Check class snowprofile}
\usage{
is.snowprofile(x)
}
\arguments{
\item{x}{object to test}
}
\value{
boolean
}
\description{
Check if object is of class \link{snowprofile}
}
