% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowprofileLayers.R
\name{is.snowprofileLayers}
\alias{is.snowprofileLayers}
\title{Check class snowprofileLayers}
\usage{
is.snowprofileLayers(x)
}
\arguments{
\item{x}{object to test}
}
\value{
boolean
}
\description{
Check if object is of class snowprofileLayers
}
