% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowprofileSet.R
\name{is.snowprofileSet}
\alias{is.snowprofileSet}
\title{Check class snowprofileSet}
\usage{
is.snowprofileSet(x)
}
\arguments{
\item{x}{object to test}
}
\value{
boolean
}
\description{
Check if object is of class \link{snowprofileSet}
}
