% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowprofileLayers.R
\name{new_snowprofileLayers}
\alias{new_snowprofileLayers}
\title{Low-level constructor for a snowprofileLayers object}
\usage{
new_snowprofileLayers(...)
}
\arguments{
\item{...}{see \link{snowprofileLayers}}
}
\value{
snowprofileLayers object as data.frame with strings as factors
}
\description{
Low-cost, efficient constructor function to be used by users who know what they're doing. \cr \cr
\strong{Important:} Make sure the last row of the data.frame corresponds to the snow surface. No checks incorporated for this low-level constructor!
}
