% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowprofile.R
\name{print.snowprofile}
\alias{print.snowprofile}
\title{Print snowprofile object}
\usage{
\method{print}{snowprofile}(x, pretty = TRUE, nLayers = NA, ...)
}
\arguments{
\item{x}{\link{snowprofile} object}

\item{pretty}{pretty print the object (data.frame-like instead of list-like)}

\item{nLayers}{only print the first few layers (cf., \link{head})}

\item{...}{passed to \link{print.default}}
}
\value{
object gets printed to console
}
\description{
Print snowprofile object
}
\examples{

## pretty print
SPpairs$A_manual
## or alternatively:
print(SPpairs$A_manual)
## reduce number of layers printed:
print(SPpairs$A_manual, nLayers = 6)

## print profile non-pretty (i.e., like the data is stored):
print(SPpairs$A_manual, pretty = FALSE)

}
