% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind.snowprofile.R
\name{rbind.snowprofile}
\alias{rbind.snowprofile}
\title{Convert snowprofile into data.frame with columns for metadata}
\usage{
\method{rbind}{snowprofile}(..., deparse.level = 1)
}
\arguments{
\item{...}{Object of class \link{snowprofile}}

\item{deparse.level}{Argument for generic rbind method}
}
\value{
data.frame
}
\description{
Convert snowprofile object into data.frame with a row for each layer and additional columns with metadata
}
\details{
Metadata columns are calculated with \link{summary.snowprofile}
}
\examples{

Profile <- SPgroup[[1]]
ProfileTable <- rbind(Profile)
head(ProfileTable)

}
\seealso{
\link{summary.snowprofile}, \link{rbind.snowprofileSet}
}
\author{
shorton
}
