% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSmet.R
\name{readSmet}
\alias{readSmet}
\title{Parse a SMET file}
\usage{
readSmet(Filename)
}
\arguments{
\item{Filename}{Path to a smet file}
}
\value{
List containing metadata and data
}
\description{
Read contents of a SMET file https://models.slf.ch/docserver/meteoio/SMET_specifications.pdf
}
\examples{

## Path to example smet
Filename <- system.file('extdata', 'example.smet', package = 'sarp.snowprofile')
Wx = readSmet(Filename)
str(Wx)

}
\seealso{
\link{writeSmet}, \link{snowprofileSno}, \link{snowprofilePrf}, \link{snowprofilePro}
}
\author{
shorton
}
