% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarp.snowprofile-package.R
\docType{package}
\name{sarp.snowprofile-package}
\alias{sarp.snowprofile}
\alias{sarp.snowprofile-package}
\title{sarp.snowprofile: Snow Profile Analysis for Snowpack and Avalanche Research}
\description{
Analysis and plotting tools for snow profile data produced from manual snowpack observations and physical snowpack models. The functions in this package support snowpack and avalanche research by reading various formats of data (including CAAML, SMET, generic csv, and outputs from the snow cover model SNOWPACK), manipulate the data, and produce graphics such as stratigraphy and time series profiles. Package developed by the Simon Fraser University Avalanche Research Program \url{http://www.avalancheresearch.ca}. Graphics apply visualization concepts from Horton, Nowak, and Haegeli (2020, \doi{10.5194/nhess-20-1557-2020}).
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.avalancheresearch.ca}
}

}
\author{
\strong{Maintainer}: Pascal Haegeli \email{pascal_haegeli@sfu.ca}

Authors:
\itemize{
  \item Simon Horton \email{shorton@sfu.ca}
  \item Florian Herla \email{fherla@sfu.ca}
}

Other contributors:
\itemize{
  \item SFU Avalanche Research Program [funder]
}

}
\keyword{internal}
