% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowprofileSet.R
\name{snowprofileSet}
\alias{snowprofileSet}
\title{Constructor for class snowprofileSet}
\usage{
snowprofileSet(x = list())
}
\arguments{
\item{x}{list of \link{snowprofile} objects}
}
\value{
a snowprofileSet
}
\description{
Constructor for class snowprofileSet
}
\seealso{
\link{snowprofile}, \link{summary.snowprofileSet}
}
