% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-SPmalformatted.R
\docType{data}
\name{SPmalformatted}
\alias{SPmalformatted}
\title{Malformatted example profiles}
\format{
A list with several entries, that are of class \link{snowprofile}
}
\usage{
SPmalformatted
}
\description{
A list with two entries, each containing a snowprofile object. Both are malformatted, check out the examples in
\link{validate_snowprofile} and \link{reformat_snowprofile} to learn how to fix it.
}
\seealso{
\link{validate_snowprofile}, \link{reformat_snowprofile}, \link{SPpairs}, \link{SPgroup}, \link{SPtimeline}
}
\keyword{object}
\keyword{snowprofile}
