% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-SPtimeline.R
\docType{data}
\name{SPtimeline}
\alias{SPtimeline}
\title{Timeseries of snowprofiles
#'}
\format{
A list with several entries, that are of class \link{snowprofile}
}
\usage{
SPtimeline
}
\description{
Timeseries of snowprofiles
#'
}
\examples{

summary(SPtimeline)
plot(SPtimeline)
}
\seealso{
\link{SPgroup}, \link{SPpairs}
}
\keyword{snowprofiles}
