% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowprofileSet.R
\name{[.snowprofileSet}
\alias{[.snowprofileSet}
\title{Extract method}
\usage{
\method{[}{snowprofileSet}(x, i)
}
\arguments{
\item{x}{object from which to extract element(s) or in which to replace element(s).}

\item{i}{indices specifying elements to extract or replace}
}
\value{
\link{snowprofileSet} object
}
\description{
Extract method
}
