% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-swisscode.R
\docType{data}
\name{swisscode}
\alias{swisscode}
\title{Numerical, Swiss Grain Type Code}
\format{
A character array
}
\usage{
swisscode
}
\description{
A character array of grain types that can be translated into a numerical code by their indices.
}
\examples{
print(swisscode)

## see numerical code for each grain type:
rbind(swisscode, seq(length(swisscode)))
}
\keyword{swisscode}
