% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowprofileLayers.R
\name{validate_snowprofileLayers}
\alias{validate_snowprofileLayers}
\title{Validate correctness of snowprofileLayers object}
\usage{
validate_snowprofileLayers(object, silent = FALSE)
}
\arguments{
\item{object}{to be tested}

\item{silent}{remain silent upon error (i.e., don't throw error, but only print it)}
}
\value{
Per default an error is raised when discovered, if \code{silent = TRUE} the error is only printed and the
error message returned.
}
\description{
Validator function that checks if class standards are being met and raises an error if not.
}
